Eliminate duplication in earthfile2llb vs earthfile2llb2.
