+++ builder/builder.go
@@ -159,37 +159,20 @@ func (b *Builder) convertAndBuild(ctx context.Context, target domain.Target, opt
 	bf := func(childCtx context.Context, gwClient gwclient.Client) (*gwclient.Result, error) {
 		var err error
 		if !b.builtMain {
-			if b.opt.EnableAst {
-				mts, err = earthfile2llb.Earthfile2LLB2(childCtx, target, earthfile2llb.ConvertOpt{
-					GwClient:             gwClient,
-					Resolver:             b.resolver,
-					ImageResolveMode:     b.opt.ImageResolveMode,
-					DockerBuilderFun:     b.MakeImageAsTarBuilderFun(),
-					CleanCollection:      b.opt.CleanCollection,
-					Platform:             opt.Platform,
-					VarCollection:        b.opt.VarCollection,
-					BuildContextProvider: b.opt.BuildContextProvider,
-					CacheImports:         b.opt.CacheImports,
-					UseInlineCache:       b.opt.UseInlineCache,
-					UseFakeDep:           b.opt.UseFakeDep,
-					EnableAst:            b.opt.EnableAst,
-				})
-			} else {
-				mts, err = earthfile2llb.Earthfile2LLB(childCtx, target, earthfile2llb.ConvertOpt{
-					GwClient:             gwClient,
-					Resolver:             b.resolver,
-					ImageResolveMode:     b.opt.ImageResolveMode,
-					DockerBuilderFun:     b.MakeImageAsTarBuilderFun(),
-					CleanCollection:      b.opt.CleanCollection,
-					Platform:             opt.Platform,
-					VarCollection:        b.opt.VarCollection,
-					BuildContextProvider: b.opt.BuildContextProvider,
-					CacheImports:         b.opt.CacheImports,
-					UseInlineCache:       b.opt.UseInlineCache,
-					UseFakeDep:           b.opt.UseFakeDep,
-					EnableAst:            b.opt.EnableAst,
-				})
-			}
+			mts, err = earthfile2llb.Earthfile2LLB(childCtx, target, earthfile2llb.ConvertOpt{
+				GwClient:             gwClient,
+				Resolver:             b.resolver,
+				ImageResolveMode:     b.opt.ImageResolveMode,
+				DockerBuilderFun:     b.MakeImageAsTarBuilderFun(),
+				CleanCollection:      b.opt.CleanCollection,
+				Platform:             opt.Platform,
+				VarCollection:        b.opt.VarCollection,
+				BuildContextProvider: b.opt.BuildContextProvider,
+				CacheImports:         b.opt.CacheImports,
+				UseInlineCache:       b.opt.UseInlineCache,
+				UseFakeDep:           b.opt.UseFakeDep,
+				EnableAst:            b.opt.EnableAst,
+			})
 			if err != nil {
 				return nil, err
 			}
